Update Main.rb
Forgot to comment with the last commit, so this commit just holds the comment for it. I have been working to debug my initialization process. I can initialize my class with a new account, but havent been able to do it with a "sign in account". Could be because it doesnt have any users stored. Will attempt a test with a dummy user entered and see if it can find it that way.